+++ src/app/pages/auth/welcome/welcome.page.ts
