--- src/app/pages/main/categories/categories-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { CategoriesPage } from './categories.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: CategoriesPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class CategoriesPageRoutingModule {}
