--- src/app/pages/main/categories/categories.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-
-import { IonicModule } from '@ionic/angular';
-
-import { CategoriesPageRoutingModule } from './categories-routing.module';
-
-import { CategoriesPage } from './categories.page';
-import { SharedModule } from "../../../shared/shared.module";
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-    CategoriesPageRoutingModule,
-    SharedModule
-],
-  declarations: [CategoriesPage]
-})
-export class CategoriesPageModule {}
