--- src/app/pages/main/categories/categories.page.html
@@ -1,40 +1,0 @@
-<app-header [title]="'Categorias'" [showBackButton]="true"></app-header>
-
-<ion-content [fullscreen]="true">
-  <div class="category-container">
-    <div class="category">
-      <img src="https://picsum.photos/600/400?random=1" alt="Imagen del servicio">
-      <p>Nom. Cat. 1</p>
-    </div>
-  </div>
-  <div class="category-container">
-    <div class="category">
-      <img src="https://picsum.photos/600/400?random=2" alt="Imagen del servicio">
-      <p>Nom. Cat. 2</p>
-    </div>
-  </div>
-  <div class="category-container">
-    <div class="category">
-      <img src="https://picsum.photos/600/400?random=3" alt="Imagen del servicio">
-      <p>Nom. Cat. 3</p>
-    </div>
-  </div>
-  <div class="category-container">
-    <div class="category">
-      <img src="https://picsum.photos/600/400?random=4" alt="Imagen del servicio">
-      <p>Nom. Cat. 4</p>
-    </div>
-  </div>
-  <div class="category-container">
-    <div class="category">
-      <img src="https://picsum.photos/600/400?random=5" alt="Imagen del servicio">
-      <p>Nom. Cat. 5</p>
-    </div>
-  </div>
-  <div class="category-container">
-    <div class="category">
-      <img src="https://picsum.photos/600/400?random=6" alt="Imagen del servicio">
-      <p>Nom. Cat. 6</p>
-    </div>
-  </div>
-</ion-content>
