--- src/app/pages/main/categories/categories.page.scss
@@ -1,36 +1,0 @@
-
-// SELECTED CATEGORY
-.category-container {
-    display: flex;
-    padding: 0px 15px;
-    position: relative;
-  }
-  
-  .category {
-    width: 100%;
-    height: 200px;
-    margin: 10px 0;
-    overflow: hidden;
-    background-color: rgba(0, 0, 0, 1);
-    border-radius: 10px;
-    
-  }
-  
-  .category img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    border-radius: 10px;
-  }
-  
-  .category p {
-    position: absolute;
-    top: 50%;
-    left: 80%;
-    transform: translate(-50%, -50%);
-    color: white;
-    font-weight: bold;
-    text-align: center;
-    font-size: 30px;
-    margin: 0;
-  }
