--- src/app/pages/main/home/home.module.ts
@@ -1,23 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-
-import { IonicModule } from '@ionic/angular';
-
-import { HomePageRoutingModule } from './home-routing.module';
-
-import { HomePage } from './home.page';
-
-import { SharedModule } from 'src/app/shared/shared.module';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-    SharedModule,
-    HomePageRoutingModule
-  ],
-  declarations: [HomePage]
-})
-export class HomePageModule {}
