--- src/app/pages/main/home/home.page.scss
@@ -1,122 +1,0 @@
-
-// CATEGORY SLIDER
-.category-slider-container {
-  background-color: #fff;
-  overflow-x: auto; 
-}
-
-.category-slider {
-  display: flex;
-  overflow-x: auto;
-  padding: 10px;
-}
-
-.category-container {
-  margin-right: 10px;
-  text-align: center;
-}
-
-.category-container p {
-  font-size: 12px;
-  font-weight: bold;
-  margin-top: 2px;
-  margin-bottom: 0;
-  color: #FF7043;
-}
-
-.category-circle {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 80px;
-  height: 80px;
-  background-color: #FF7043;
-  color: white;
-  border-radius: 50%;
-
-}
-
-.category-circle ion-avatar {
-  width: 75px;
-  height: 75px;
-}
-
-// SELECTED CATEGORY
-.selected-category-container {
-  display: flex;
-  padding: 0px 15px;
-  position: relative;
-}
-
-.selected-category {
-  width: 100%;
-  height: 200px;
-  margin: 10px 0;
-  overflow: hidden;
-  background-color: rgba(0, 0, 0, 1);
-  border-radius: 10px;
-  
-}
-
-.selected-category img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  border-radius: 10px;
-}
-
-.selected-category p {
-  position: absolute;
-  top: 50%;
-  left: 80%;
-  transform: translate(-50%, -50%);
-  color: white;
-  font-weight: bold;
-  text-align: center;
-  font-size: 30px;
-  margin: 0;
-}
-
-/* General */
-ion-content {
-  --background: #f6f6f6;
-}
-
-
-/* Estilo para las imágenes destacadas */
-.image-carousel {
-  width: 100%;
-  height: 200px;
-  margin: 10px 0;
-  overflow: hidden;
-}
-
-.carousel-container {
-  display: flex;
-  gap: 10px;
-  transition: transform 0.3s ease-in-out;
-}
-
-.carousel-image img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  border-radius: 10px;
-}
-
-/* Lista de servicios */
-.service-list {
-  margin: 10px;
-}
-
-.service-list ion-card {
-  --border-radius: 12px;
-  margin-bottom: 15px;
-}
-
-.service-list ion-avatar {
-  width: 60px;
-  height: 60px;
-  border-radius: 12px;
-}
