--- src/app/pages/main/main-routing.module.ts
@@ -1,39 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { MainPage } from './main.page';
-import { LobbyPage } from '../lobby/lobby.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: LobbyPage
-  },
-  {
-    path: 'home',
-    loadChildren: () => import('./home/home.module').then( m => m.HomePageModule)
-  },
-  {
-    path: 'profile',
-    loadChildren: () => import('./profile/profile.module').then( m => m.ProfilePageModule)
-  },
-  {
-    path: 'welcome',
-    loadChildren: () => import('./welcome/welcome.module').then( m => m.WelcomePageModule)
-  },
-  {
-    path: 'categories',
-    loadChildren: () => import('./categories/categories.module').then( m => m.CategoriesPageModule)
-  },
-  {
-    path: 'lobby',
-    loadChildren: () => import('../lobby/lobby.module').then( m => m.LobbyPageModule)
-  }
-
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class MainPageRoutingModule {}
