--- src/app/pages/main/main.page.html
@@ -1,18 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-title>main</ion-title>
-  </ion-toolbar>
-</ion-header>
-
-
-<ion-buttons>
-  <ion-back-button *ngIf="backButton"   [defaultHref]="backButton"></ion-back-button>
-</ion-buttons>
-
-<ion-content [fullscreen]="true">
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title size="large">main</ion-title>
-    </ion-toolbar>
-  </ion-header>
-</ion-content>
