--- src/app/pages/main/main.page.ts
@@ -1,18 +1,0 @@
-import { Component, Input, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-main',
-  templateUrl: './main.page.html',
-  styleUrls: ['./main.page.scss'],
-})
-export class MainPage implements OnInit {
-
-
-  @Input() backButton!: string; 
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
