--- src/app/pages/main/profile/profile-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { ProfilePage } from './profile.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: ProfilePage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class ProfilePageRoutingModule {}
