--- src/app/pages/main/profile/profile.page.html
@@ -1,75 +1,0 @@
-<app-header [title]="'Perfil'" [showBackButton]="true"></app-header>
-
-<ion-content [fullscreen]="true">
-
-
-  <ion-tabs>
-    <ion-tab-bar slot="bottom">
-      <ion-tab-button (click)="home()">
-        <ion-icon name="home-outline"></ion-icon>
-        Inicio
-      </ion-tab-button>
-      <ion-tab-button (click)="Profile()">
-        <ion-icon name="Person-outline"></ion-icon>
-        Perfil
-      </ion-tab-button>
-    </ion-tab-bar>
-  </ion-tabs>
-
-  <div class="info-container">
-    <h2>Información de la cuenta</h2>
-
-    <div class="profile-container">
-      <img [src]="profileImage ? profileImage : 'assets/default-profile.png'" class="profile-image" alt="Imagen de perfil"/>
-      
-      <!-- Input de tipo file oculto -->
-      <input type="file" #fileInput hidden (change)="onFileSelected($event)" accept="image/*" />
-
-      <!-- Botón para cambiar la foto -->
-      <ion-button class="custom-button" (click)="triggerFileInput()">Cambiar Foto</ion-button>
-    </div>
-
-    <h2>Información básica</h2>
-
-    <ion-list lines="none">
-
-      <!-- Nombre de Usuario -->
-      <ion-item>
-        <ion-label>
-          <h2>Nombre</h2>
-          <p>{{ userData?.name }}</p>
-        </ion-label>
-      </ion-item>
-
-      <!-- Correo Electrónico -->
-      <ion-item>
-        <ion-label>
-          <h2>Correo Electrónico</h2>
-          <p>{{ userData?.email }}</p>
-        </ion-label>
-      </ion-item>
-
-      <!-- Número de Contacto -->
-      <ion-item>
-        <ion-label>
-          <h2>Número de contacto</h2>
-          <p>{{ userData?.phone }}</p>
-        </ion-label>
-      </ion-item>
-
-      <!-- Ubicación -->
-      <ion-item>
-        <ion-label>
-          <h2>Ubicación</h2>
-          <p>{{ userData?.location }}</p>
-        </ion-label>
-      </ion-item>
-
-    </ion-list>
-  </div>
-
-
-</ion-content>
-
-  <!-- Contenido de la página -->
-  <app-bottom-nav></app-bottom-nav> <!-- Añadir la barra de navegación inferior -->
