--- src/app/pages/main/welcome/welcome.module.ts
@@ -1,18 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-import { IonicModule } from '@ionic/angular';
-
-import { WelcomePageRoutingModule } from './welcome-routing.module';
-import { WelcomePage } from './welcome.page';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-    WelcomePageRoutingModule
-  ],
-  declarations: [WelcomePage]
-})
-export class WelcomePageModule {}
